improved ui and folder restructuring
