--- front-end/src/components/Elements/Circle/Circle.css
@@ -1,3 +1,0 @@
-.Circle-main{
-    color: red; 
-}
