--- front-end/src/components/Elements/Circle/Circle.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import "./Circle.css" 
-
-function Circle() {
-  return (
-    <div className="Circle-main">This is a circle component</div>
-  )
-}
-
-export default Circle
