+++ front-end/src/components/Elements/MangaIcon/MangaIcon.css
@@ -2,9 +2,7 @@
     display:flex; 
     flex-direction:column; 
     align-items:center; 
-    max-width: 25%; 
-    padding: 20px; 
-    margin: 10px; 
+    margin: 4%; 
 }
 
 .MangaIcon-main img{
