+++ front-end/src/components/Elements/MangaInfo/MangaInfo.css
@@ -1,0 +1,7 @@
+.MangaInfo-main{
+
+}
+
+.MangaInfo-chatper-tracker{
+
+}
