+++ front-end/src/components/Elements/index.js
@@ -1,4 +1,14 @@
-export * from "./Circle"
+//manga image
 export * from "./MangaIcon"
+
+//star prop 
 export * from "./Star"
+
+//remove or add users in profile page
 export * from "./UserItem"
+
+//manga information
+export * from "./MangaInfo" 
+
+//Idk what this is help
+export * from "./MangeProfileImage"
