+++ front-end/src/components/Layout/Header/Header.css
@@ -171,7 +171,7 @@ i {
 
 /* Connecting to Outlet which is reactjs component */
 #detail {
-    flex: 1;
+    /* flex: 1; */
     padding: 2rem 4rem;
     width: 75%;
 }
