+++ front-end/src/components/Layout/MangaRow/MangaRow.css
@@ -2,11 +2,9 @@
     display: flex; 
     flex-direction: row;
     flex-wrap: nowrap;  
-    overflow-x: auto; 
     overflow: auto; 
-    min-width: 50vw; 
-    max-width: 100vw; 
-
+    
+    
     -ms-overflow-style: none;
     /* IE and Edge */
     scrollbar-width: none;
