+++ front-end/src/components/index.js
@@ -1,0 +1,5 @@
+//Elements are basic components (clickable, icons, etc)
+export * from Elements
+
+//Layout encompass more than elements but not enough to be a whole page (row of clickable, row of icons, tables, etc)
+export * from Layout
