+++ front-end/src/pages/Works/Works.css
@@ -1,0 +1,5 @@
+.Works-main{
+    height: 100%;
+    overflow-y: auto; 
+    overflow-x: hidden; 
+}
