+++ front-end/src/pages/index.js
@@ -1,0 +1,20 @@
+//Login page 
+export * from Auth
+
+//Error page for client side error
+export * from Error
+
+//Following page 
+export * from Follow
+
+//Dashboard page 
+export * from Home
+
+//Idk what this is someone fill it in
+export * from MangaWorks
+
+//User profile page 
+export * from Profile
+
+//Works page represent page of each individual manga
+export * from Works
